+++ src/mlpack/methods/random_forest/random_forest.hpp
@@ -51,7 +51,7 @@ class RandomForest
    * Construct the random forest without any training or specifying the number
    * of trees.  Predict() will throw an exception until Train() is called.
    */
-  RandomForest() { }
+  RandomForest();
 
   /**
    * Create a random forest, training on the given labeled training data with
@@ -183,6 +183,8 @@ class RandomForest
    * @param minimumLeafSize Minimum number of points in each tree's leaf nodes.
    * @param minimumGainSplit Minimum gain for splitting a decision tree node.
    * @param maximumDepth Maximum depth for the tree.
+   * @param warmStart When set to `true`, it adds `numTrees` new trees to the
+   *     existing random forest otherwise a new forest is trained from scratch.
    * @param dimensionSelector Instantiated dimension selection policy.
    * @return The average entropy of all the decision trees trained under forest.
    */
@@ -194,6 +196,7 @@ class RandomForest
                const size_t minimumLeafSize = 1,
                const double minimumGainSplit = 1e-7,
                const size_t maximumDepth = 0,
+               const bool warmStart = false,
                DimensionSelectionType dimensionSelector =
                    DimensionSelectionType());
 
@@ -214,6 +217,8 @@ class RandomForest
    * @param minimumLeafSize Minimum number of points in each tree's leaf nodes.
    * @param minimumGainSplit Minimum gain for splitting a decision tree node.
    * @param maximumDepth Maximum depth for the tree.
+   * @param warmStart When set to `true`, it adds `numTrees` new trees to the
+   *     existing random forest else a new forest is trained from scratch.
    * @param dimensionSelector Instantiated dimension selection policy.
    * @return The average entropy of all the decision trees trained under forest.
    */
@@ -226,6 +231,7 @@ class RandomForest
                const size_t minimumLeafSize = 1,
                const double minimumGainSplit = 1e-7,
                const size_t maximumDepth = 0,
+               const bool warmStart = false,
                DimensionSelectionType dimensionSelector =
                    DimensionSelectionType());
 
@@ -244,6 +250,8 @@ class RandomForest
    * @param minimumLeafSize Minimum number of points in each tree's leaf nodes.
    * @param minimumGainSplit Minimum gain for splitting a decision tree node.
    * @param maximumDepth Maximum depth for the tree.
+   * @param warmStart When set to `true`, it adds `numTrees` new trees to the
+   *     existing random forest else a new forest is trained from scratch.
    * @param dimensionSelector Instantiated dimension selection policy.
    * @return The average entropy of all the decision trees trained under forest.
    */
@@ -256,6 +264,7 @@ class RandomForest
                const size_t minimumLeafSize = 1,
                const double minimumGainSplit = 1e-7,
                const size_t maximumDepth = 0,
+               const bool warmStart = false,
                DimensionSelectionType dimensionSelector =
                    DimensionSelectionType());
 
@@ -276,6 +285,8 @@ class RandomForest
    * @param minimumLeafSize Minimum number of points in each tree's leaf nodes.
    * @param minimumGainSplit Minimum gain for splitting a decision tree node.
    * @param maximumDepth Maximum depth for the tree.
+   * @param warmStart When set to `true`, it adds `numTrees` new trees to the
+   *     existing random forest else a new forest is trained from scratch.
    * @param dimensionSelector Instantiated dimension selection policy.
    * @return The average entropy of all the decision trees trained under forest.
    */
@@ -289,6 +300,7 @@ class RandomForest
                const size_t minimumLeafSize = 1,
                const double minimumGainSplit = 1e-7,
                const size_t maximumDepth = 0,
+               const bool warmStart = false,
                DimensionSelectionType dimensionSelector =
                    DimensionSelectionType());
 
@@ -370,6 +382,8 @@ class RandomForest
    * @param minimumGainSplit Minimum gain for splitting a decision tree node.
    * @param maximumDepth Maximum depth for the tree.
    * @param dimensionSelector Instantiated dimension selection policy.
+   * @param warmStart When set to `true`, it fits new trees and add them to the
+   *     previous forest else a new forest is trained from scratch.
    * @tparam UseWeights Whether or not to use the weights parameter.
    * @tparam UseDatasetInfo Whether or not to use the datasetInfo parameter.
    * @tparam MatType The type of data matrix (i.e. arma::mat).
@@ -385,10 +399,14 @@ class RandomForest
                const size_t minimumLeafSize,
                const double minimumGainSplit,
                const size_t maximumDepth,
-               DimensionSelectionType& dimensionSelector);
+               DimensionSelectionType& dimensionSelector,
+               const bool warmStart = false);
 
   //! The trees in the forest.
   std::vector<DecisionTreeType> trees;
+
+  //! The average gain of the forest.
+  double avgGain;
 };
 
 } // namespace tree