+++ src/mlpack/methods/random_forest/random_forest_impl.hpp
@@ -24,6 +24,23 @@ template<
     template<typename> class NumericSplitType,
     template<typename> class CategoricalSplitType
 >
+RandomForest<
+    FitnessFunction,
+    DimensionSelectionType,
+    NumericSplitType,
+    CategoricalSplitType
+>::RandomForest() :
+    avgGain(0.0)
+{
+  // Nothing to do here.
+}
+
+template<
+    typename FitnessFunction,
+    typename DimensionSelectionType,
+    template<typename> class NumericSplitType,
+    template<typename> class CategoricalSplitType
+>
 template<typename MatType>
 RandomForest<
     FitnessFunction,
@@ -37,13 +54,15 @@ RandomForest<
                 const size_t minimumLeafSize,
                 const double minimumGainSplit,
                 const size_t maximumDepth,
-                DimensionSelectionType dimensionSelector)
+                DimensionSelectionType dimensionSelector) :
+    avgGain(0.0)
 {
   // Pass off work to the Train() method.
   data::DatasetInfo info; // Ignored.
   arma::rowvec weights; // Fake weights, not used.
   Train<false, false>(dataset, info, labels, numClasses, weights, numTrees,
-      minimumLeafSize, minimumGainSplit, maximumDepth, dimensionSelector);
+      minimumLeafSize, minimumGainSplit, maximumDepth, dimensionSelector,
+      false);
 }
 
 template<
@@ -66,13 +85,14 @@ RandomForest<
                 const size_t minimumLeafSize,
                 const double minimumGainSplit,
                 const size_t maximumDepth,
-                DimensionSelectionType dimensionSelector)
+                DimensionSelectionType dimensionSelector):
+                    avgGain(0.0)
 {
   // Pass off work to the Train() method.
   arma::rowvec weights; // Fake weights, not used.
   Train<false, true>(dataset, datasetInfo, labels, numClasses, weights,
       numTrees, minimumLeafSize, minimumGainSplit, maximumDepth,
-      dimensionSelector);
+      dimensionSelector, false);
 }
 
 template<
@@ -95,12 +115,14 @@ RandomForest<
                 const size_t minimumLeafSize,
                 const double minimumGainSplit,
                 const size_t maximumDepth,
-                DimensionSelectionType dimensionSelector)
+                DimensionSelectionType dimensionSelector) :
+    avgGain(0.0)
 {
   // Pass off work to the Train() method.
   data::DatasetInfo info; // Ignored by Train().
   Train<true, false>(dataset, info, labels, numClasses, weights, numTrees,
-      minimumLeafSize, minimumGainSplit, maximumDepth, dimensionSelector);
+      minimumLeafSize, minimumGainSplit, maximumDepth, dimensionSelector,
+      false);
 }
 
 template<
@@ -124,11 +146,13 @@ RandomForest<
                 const size_t minimumLeafSize,
                 const double minimumGainSplit,
                 const size_t maximumDepth,
-                DimensionSelectionType dimensionSelector)
+                DimensionSelectionType dimensionSelector) :
+    avgGain(0.0)
 {
   // Pass off work to the Train() method.
-  Train<true, true>(dataset, datasetInfo, labels, numClasses, weights, numTrees,
-      minimumLeafSize, minimumGainSplit, maximumDepth, dimensionSelector);
+  Train<true, true>(dataset, datasetInfo, labels, numClasses, weights,
+      numTrees, minimumLeafSize, minimumGainSplit, maximumDepth,
+      dimensionSelector, false);
 }
 
 template<
@@ -150,14 +174,15 @@ double RandomForest<
          const size_t minimumLeafSize,
          const double minimumGainSplit,
          const size_t maximumDepth,
+         const bool warmStart,
          DimensionSelectionType dimensionSelector)
 {
   // Pass off to Train().
-  data::DatasetInfo info; // Ignored by Train().
+  data::DatasetInfo datasetInfo; // Ignored by Train().
   arma::rowvec weights; // Ignored by Train().
-  return Train<false, false>(dataset, info, labels, numClasses, weights,
+  return Train<false, false>(dataset, datasetInfo, labels, numClasses, weights,
       numTrees, minimumLeafSize, minimumGainSplit, maximumDepth,
-      dimensionSelector);
+      dimensionSelector, warmStart);
 }
 
 template<
@@ -180,13 +205,14 @@ double RandomForest<
          const size_t minimumLeafSize,
          const double minimumGainSplit,
          const size_t maximumDepth,
+         const bool warmStart,
          DimensionSelectionType dimensionSelector)
 {
   // Pass off to Train().
   arma::rowvec weights; // Ignored by Train().
   return Train<false, true>(dataset, datasetInfo, labels, numClasses, weights,
       numTrees, minimumLeafSize, minimumGainSplit, maximumDepth,
-      dimensionSelector);
+      dimensionSelector, warmStart);
 }
 
 template<
@@ -209,13 +235,14 @@ double RandomForest<
          const size_t minimumLeafSize,
          const double minimumGainSplit,
          const size_t maximumDepth,
+         const bool warmStart,
          DimensionSelectionType dimensionSelector)
 {
   // Pass off to Train().
-  data::DatasetInfo info; // Ignored by Train().
-  return Train<false, false>(dataset, info, labels, numClasses, weights,
+  data::DatasetInfo datasetInfo; // Ignored by Train().
+  return Train<false, false>(dataset, datasetInfo, labels, numClasses, weights,
       numTrees, minimumLeafSize, minimumGainSplit, maximumDepth,
-      dimensionSelector);
+      dimensionSelector, warmStart);
 }
 
 template<
@@ -239,12 +266,13 @@ double RandomForest<
          const size_t minimumLeafSize,
          const double minimumGainSplit,
          const size_t maximumDepth,
+         const bool warmStart,
          DimensionSelectionType dimensionSelector)
 {
   // Pass off to Train().
   return Train<true, true>(dataset, datasetInfo, labels, numClasses, weights,
       numTrees, minimumLeafSize, minimumGainSplit, maximumDepth,
-      dimensionSelector);
+      dimensionSelector, warmStart);
 }
 
 template<
@@ -410,6 +438,7 @@ void RandomForest<
     trees.resize(numTrees);
 
   ar(CEREAL_NVP(trees));
+  ar(CEREAL_NVP(avgGain));
 }
 
 template<
@@ -433,13 +462,20 @@ double RandomForest<
          const size_t minimumLeafSize,
          const double minimumGainSplit,
          const size_t maximumDepth,
-         DimensionSelectionType& dimensionSelector)
+         DimensionSelectionType& dimensionSelector,
+         const bool warmStart)
 {
-  // Train each tree individually.
-  trees.resize(numTrees); // This will fill the vector with untrained trees.
-  double avgGain = 0.0;
+  // Reset the forest if we are not doing a warm-start.
+  if (!warmStart)
+    trees.clear();
+  const size_t oldNumTrees = trees.size();
+  trees.resize(trees.size() + numTrees);
 
-  #pragma omp parallel for reduction( + : avgGain)
+  // Convert avgGain to total gain.
+  double totalGain = avgGain * oldNumTrees;
+
+  // Train each tree individually.
+  #pragma omp parallel for reduction( + : totalGain)
   for (omp_size_t i = 0; i < numTrees; ++i)
   {
     Timer::Start("bootstrap");
@@ -450,40 +486,44 @@ double RandomForest<
         bootstrapLabels, bootstrapWeights);
     Timer::Stop("bootstrap");
 
-    // Now build the decision tree.
     Timer::Start("train_tree");
     if (UseWeights)
     {
       if (UseDatasetInfo)
       {
-        avgGain += trees[i].Train(bootstrapDataset, datasetInfo,
-            bootstrapLabels, numClasses, bootstrapWeights, minimumLeafSize,
-            minimumGainSplit, maximumDepth, dimensionSelector);
+        totalGain += trees[oldNumTrees + i].Train(bootstrapDataset,
+            datasetInfo, bootstrapLabels, numClasses, bootstrapWeights,
+            minimumLeafSize, minimumGainSplit, maximumDepth,
+            dimensionSelector);
       }
       else
       {
-        avgGain += trees[i].Train(bootstrapDataset, bootstrapLabels, numClasses,
-            bootstrapWeights, minimumLeafSize, minimumGainSplit, maximumDepth,
-            dimensionSelector);
+        totalGain += trees[oldNumTrees + i].Train(bootstrapDataset,
+            bootstrapLabels, numClasses, bootstrapWeights, minimumLeafSize,
+            minimumGainSplit, maximumDepth, dimensionSelector);
       }
     }
     else
     {
       if (UseDatasetInfo)
       {
-        avgGain += trees[i].Train(bootstrapDataset, datasetInfo,
-            bootstrapLabels, numClasses, minimumLeafSize, minimumGainSplit,
-            maximumDepth, dimensionSelector);
+        totalGain += trees[oldNumTrees + i].Train(bootstrapDataset,
+            datasetInfo, bootstrapLabels, numClasses, minimumLeafSize,
+            minimumGainSplit, maximumDepth, dimensionSelector);
       }
       else
       {
-        avgGain += trees[i].Train(bootstrapDataset, bootstrapLabels, numClasses,
-            minimumLeafSize, minimumGainSplit, maximumDepth, dimensionSelector);
+        totalGain += trees[oldNumTrees + i].Train(bootstrapDataset,
+            bootstrapLabels, numClasses, minimumLeafSize, minimumGainSplit,
+            maximumDepth, dimensionSelector);
       }
     }
+
     Timer::Stop("train_tree");
   }
-  return avgGain / numTrees;
+
+  avgGain = totalGain / trees.size();
+  return avgGain;
 }
 
 } // namespace tree