+++ src/mlpack/methods/random_forest/random_forest_main.cpp
@@ -130,6 +130,8 @@ PARAM_INT_IN("subspace_dim", "Dimensionality of random subspace to use for "
     "d", 0);
 
 PARAM_INT_IN("seed", "Random seed.  If 0, 'std::time(NULL)' is used.", "s", 0);
+PARAM_FLAG("warm_start", "If true and passed along with `training` and "
+    "`input_model` then trains more trees on top of existing model.", "w");
 
 /**
  * This is the class that we will serialize.  It is a pretty simple wrapper
@@ -167,7 +169,11 @@ static void mlpackMain()
     math::RandomSeed((size_t) std::time(NULL));
 
   // Check for incompatible input parameters.
-  RequireOnlyOnePassed({ "training", "input_model" }, true);
+  if (!IO::HasParam("warm_start"))
+    RequireOnlyOnePassed({ "training", "input_model" }, true);
+  else
+    // When warm_start is passed, training and input_model must also be passed.
+    RequireNoneOrAllPassed({"warm_start", "training", "input_model"}, true);
 
   ReportIgnoredParam({{ "training", false }}, "print_training_accuracy");
   ReportIgnoredParam({{ "test", false }}, "test_labels");
@@ -201,10 +207,17 @@ static void mlpackMain()
   ReportIgnoredParam({{ "training", false }}, "minimum_leaf_size");
 
   RandomForestModel* rfModel;
+  // Input model is loaded when we are either doing warm-started training or
+  // else we are making predictions only or both.
+  if (IO::HasParam("input_model"))
+    rfModel = IO::GetParam<RandomForestModel*>("input_model");
+  // Handles the case when we are training new forest from scratch.
+  else
+    rfModel = new RandomForestModel();
+
   if (IO::HasParam("training"))
   {
     Timer::Start("rf_training");
-    rfModel = new RandomForestModel();
 
     // Train the model on the given input data.
     arma::mat data = std::move(IO::GetParam<arma::mat>("training"));
@@ -233,7 +246,8 @@ static void mlpackMain()
 
     // Train the model.
     rfModel->rf.Train(data, labels, numClasses, numTrees, minimumLeafSize,
-        minimumGainSplit, maxDepth, mrds);
+        minimumGainSplit, maxDepth, IO::HasParam("warm_start"), mrds);
+
     Timer::Stop("rf_training");
 
     // Did we want training accuracy?
@@ -250,11 +264,6 @@ static void mlpackMain()
           << endl;
       Timer::Stop("rf_prediction");
     }
-  }
-  else
-  {
-    // Then we must be loading a model.
-    rfModel = IO::GetParam<RandomForestModel*>("input_model");
   }
 
   if (IO::HasParam("test"))